+++ r-data-factory.tf
@@ -33,7 +33,7 @@ resource "azurerm_data_factory" "main_data_factory" {
   }
 
   dynamic "vsts_configuration" {
-    for_each = var.vsts_configuration == null ? [] : ["_"]
+    for_each = var.azure_devops_configuration == null ? [] : ["_"]
 
     content {
       account_name    = vsts_configuration.value.account_name
@@ -46,4 +46,46 @@ resource "azurerm_data_factory" "main_data_factory" {
   }
 
   tags = merge(local.default_tags, var.extra_tags)
+}
+
+
+resource "azurerm_data_factory_integration_runtime_azure" "integration_runtime" {
+  count = var.integration_runtime_type == "Azure" ? 1 : 0
+
+  data_factory_name   = azurerm_data_factory.main_data_factory.name
+  name                = var.integration_runtime_custom_name
+  location            = var.location
+  resource_group_name = var.resource_group_name
+  description         = var.integration_runtime_description
+
+  cleanup_enabled         = lookup(var.azure_integration_runtime_configuration, "cleanup_enabled", true)
+  compute_type            = lookup(var.azure_integration_runtime_configuration, "compute_type", "General")
+  core_count              = lookup(var.azure_integration_runtime_configuration, "core_count", 8)
+  time_to_live_min        = lookup(var.azure_integration_runtime_configuration, "time_to_live_min", 0)
+  virtual_network_enabled = lookup(var.azure_integration_runtime_configuration, "virtual_network_enabled", false)
+}
+
+resource "azurerm_data_factory_integration_runtime_self_hosted" "integration_runtime" {
+  count = var.integration_runtime_type == "SelfHosted" ? 1 : 0
+
+  data_factory_name   = azurerm_data_factory.main_data_factory.name
+  name                = var.integration_runtime_custom_name
+  resource_group_name = var.resource_group_name
+  description         = var.integration_runtime_description
+}
+
+resource "azurerm_data_factory_integration_runtime_azure_ssis" "integration_runtime" {
+  count = var.integration_runtime_type == "AzureSSIS" ? 1 : 0
+
+  data_factory_name   = azurerm_data_factory.main_data_factory.name
+  name                = var.integration_runtime_custom_name
+  location            = var.location
+  resource_group_name = var.resource_group_name
+  description         = var.integration_runtime_description
+
+  node_size                        = var.azure_ssis_integration_runtime_configuration.node_size
+  number_of_nodes                  = lookup(var.azure_ssis_integration_runtime_configuration, "number_of_nodes", 1)
+  max_parallel_executions_per_node = lookup(var.azure_ssis_integration_runtime_configuration, "max_parallel_executions_per_node", 1)
+  edition                          = lookup(var.azure_ssis_integration_runtime_configuration, "edition", "Standard")
+  license_type                     = lookup(var.azure_ssis_integration_runtime_configuration, "license_type", "LicenseIncluded")
 }