+++ versions.tf
@@ -1,9 +1,9 @@
 terraform {
-  required_version = "> 0.13"
+  required_version = ">= 0.13"
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = ">= 2.68"
+      version = ">= 2.76"
     }
     azurecaf = {
       source  = "aztfmod/azurecaf"