+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased
 
 Added
+  * AZ-612: Add integration runtime resource to module
   * AZ-563: Initialize Azure Data Factory module